Split math code into multiple files
